修改统计维护人员接单bug

+++ classrepair-repair-item/src/main/java/team/weacsoft/repair/dto/response/StatisticsFromEquipment.java
@@ -15,7 +15,7 @@ public class StatisticsFromEquipment {
     /**
      * 学号/工号
      */
-    private int receiverIdentityId=0;
+    private Long receiverIdentityId=0L;
 
     /**
      * id
